Show success message after sending password reset email
