+++ app/auth/password-reset/page.tsx
@@ -1,6 +1,6 @@
 import PageContainer from '@/components/layout/pages/page-container';
 import PageHeader from '@/components/layout/pages/page-header';
-import EmailForm from '@/components/auth/password-reset/email-form';
+import EmailForm from '@/components/auth/password-reset/email-form/email-form';
 
 export default function PasswordResetPage() {
   return (
