--- components/auth/password-reset/email-form.tsx
@@ -1,42 +1,0 @@
-'use client';
-
-import { useFormState } from 'react-dom';
-import { Label, TextInput } from 'flowbite-react';
-
-import { submitEmailForPasswordReset } from '@/lib/actions';
-import FormSubmitButton from '../../ui/form-submit-button';
-
-export default function EmailForm() {
-  const [formState, formAction] = useFormState(submitEmailForPasswordReset, {
-    status: null
-  });
-
-  const { status, message } = formState;
-
-  return (
-    <div>
-      <form
-        action={formAction}
-        className="flex max-w-md flex-col gap-4 mx-auto"
-      >
-        {status === 'failure' && (
-          <p className="text-red-600 text-center mb-6">{message}</p>
-        )}
-
-        <p className="text-sm">
-          Please enter the email address associated with your account, and we'll
-          send you an email.
-        </p>
-
-        <div>
-          <div className="mb-2 block">
-            <Label htmlFor="email" value="Email" />
-          </div>
-          <TextInput id="email" name="email" type="email" required />
-        </div>
-
-        <FormSubmitButton />
-      </form>
-    </div>
-  );
-}
